feat(sln): refactor to latest Xperience, add standard files
