--- .config/dotnet-tools.json
@@ -1,12 +1,0 @@
-{
-  "version": 1,
-  "isRoot": true,
-  "tools": {
-    "azuresigntool": {
-      "version": "4.0.1",
-      "commands": [
-        "azuresigntool"
-      ]
-    }
-  }
-}
