+++ .gitignore
@@ -6,11 +6,14 @@
 
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
-
+##
+## Get latest from https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
 
 # User-specific files
 *.suo
-*.user
+*.csproj.user
+*.ReSharper.user
+*.DotSettings.user
 *.sln.docstates
 .vs
 
@@ -19,13 +22,9 @@
 [Dd]ebug/
 [Rr]elease/
 # x64/
-build/
 [Bb]in/
 [Oo]bj/
 [Oo]utput/
-
-# Enable "build/" folder in the NuGet Packages folder since NuGet packages use it for MSBuild targets
-!packages/*/build/
 
 # MSTest test Results
 [Tt]est[Rr]esult*/
@@ -56,66 +55,20 @@ build/
 *.scc
 *.jfm
 
-
-# Visual C++ cache files
-ipch/
-*.aps
-*.ncb
-*.opensdf
-*.sdf
-*.cachefile
-
 # Visual Studio profiler
 *.psess
 *.vsp
 *.vspx
 
-# Guidance Automation Toolkit
-*.gpState
-
 # ReSharper is a .NET coding add-in
 _ReSharper*/
 *.[Rr]e[Ss]harper
 
-# TeamCity is a build add-in
-_TeamCity*
-
 # DotCover is a Code Coverage Tool
 *.dotCover
 
-# NCrunch
-*.ncrunch*
-.*crunch*.local.xml
-
-# Installshield output folder
-[Ee]xpress/
-
-# DocProject is a documentation generator add-in
-DocProject/buildhelp/
-DocProject/Help/*.HxT
-DocProject/Help/*.HxC
-DocProject/Help/*.hhc
-DocProject/Help/*.hhk
-DocProject/Help/*.hhp
-DocProject/Help/Html2
-DocProject/Help/html
-
-# Click-Once directory
-# publish/
-
 # Publish Web Output
 *.Publish.xml
-
-# NuGet Packages Directory
-## TODO: If you have NuGet Package Restore enabled, uncomment the next line
-packages/
-
-# Windows Azure Build Output
-csx
-*.build.csdef
-
-# Windows Store app package directory
-AppPackages/
 
 # Others
 sql/
@@ -129,25 +82,9 @@ ClientBin/
 # *.pfx
 *.publishsettings
 
-# RIA/Silverlight projects
-Generated_Code/
-
-# Backup & report files from converting an old project file to a newer
-# Visual Studio version. Backup files are not needed, because we have git ;-)
-_UpgradeReport_Files/
-Backup*/
-UpgradeLog*.XML
-UpgradeLog*.htm
-
 # SQL Server files
 App_Data/*.mdf
 App_Data/*.ldf
-
-
-#LightSwitch generated files
-GeneratedArtifacts/
-_Pvt_Extensions/
-ModelManifest.xml
 
 # =========================
 # Windows detritus
@@ -169,98 +106,34 @@ $RECYCLE.BIN/
 # Web.Common junction
 */Web.*/Common/
 
-# =========================
-# Android
-# =========================
-
-# Eclipse
-.project
-.classpath
-.settings
-.checkstyle
-
-# IntelliJ IDEA
-.idea
-*.iml
-*.ipr
-*.iws
-classes
-gen-external-apklibs
-
-# Gradle
-.gradle
-build
-
-# Maven
-target
-release.properties
-pom.xml.*
-
-# Ant
-bin
-gen
-build.xml
-ant.properties
-local.properties
-proguard.cfg
-proguard-project.txt
-
-# Other
-.DS_Store
-tmp
-/sourcetree.license
-
-# =========================
-# iOS
-# =========================
-
-# Xcode
-#
-# gitignore contributors: remember to update Global/Xcode.gitignore, Objective-C.gitignore & Swift.gitignore
-
-## Build generated
-build/
-DerivedData
-
-## Various settings
-*.pbxuser
-!default.pbxuser
-*.mode1v3
-!default.mode1v3
-*.mode2v3
-!default.mode2v3
-*.perspectivev3
-!default.perspectivev3
-xcuserdata
-
-## Other
-*.xccheckout
-*.moved-aside
-*.xcuserstate
-*.xcscmblueprint
-
-## Obj-C/Swift specific
-*.hmap
-*.ipa
-
-# CocoaPods
-#
-# We recommend against adding the Pods directory to your .gitignore. However
-# you should judge for yourself, the pros and cons are mentioned at:
-# http://guides.cocoapods.org/using/using-cocoapods.html#should-i-check-the-pods-directory-into-source-control
-#
-#Pods/
-
-# Database projects - publish manifest
-#!/Deployment/**/*.Publish.xml
-
 *.jfm
 *.svclog
 *.nupkg
 
 # nodejs
 node_modules
-dist
 
-Kentico.Xperience.Contacts.Importer.sln.DotSettings
-data/*.csv
+# JetBrains Rider
+.idea/
+*.sln.iml
+
+# CodeRush personal settings
+.cr/personal
+
+#Exclude SmartSearch files
+*.fdt
+*.fdx
+*.fnm
+*.nrm
+*.prx
+*.frq
+*.tii
+*.tis
+*.del
+*.cfs
+segments_2i
+
+# Generated files
+**/Client/dist/*
+**/nuget-local/*.*pkg
+
