+++ Directory.build.targets
@@ -8,7 +8,11 @@
             <AssemblyToSign Include="$(TargetPath)" />
             <AssemblyToSign Include="$(XmlSerializersTargetPath)" Condition="Exists('$(XmlSerializersTargetPath)')" />
         </ItemGroup>
-
-        <Exec Command="dotnet AzureSignTool sign --azure-key-vault-url $(AuthenticodeKeyVaultUrl) --azure-key-vault-tenant-id $(AuthenticodeTenantId) --azure-key-vault-client-id $(AuthenticodeClientId) --azure-key-vault-client-secret $(AuthenticodeClientSecret) --azure-key-vault-certificate $(AuthenticodeCertificateName) --timestamp-rfc3161 $(TimestampServerUrl) --skip-signed %(AssemblyToSign.Identity)" />
     </Target>
+
+    <ItemGroup>
+        <Content Remove="Admin/Client/*.json" />
+        <Content Remove="Admin/*.json" />
+        <Content Remove="../Kentico.Xperience.Contacts.Importer/*.json"/>
+    </ItemGroup>
 </Project>
