+++ docs/Contributing.md
@@ -1,0 +1,12 @@
+# Contributing
+
+## Requirements
+
+- .NET SDK >= 8.0.303
+
+  - <https://dotnet.microsoft.com/en-us/download/dotnet/8.0>
+
+- Node.js >= 20.15.1
+
+  - <https://nodejs.org/en/download>
+  - <https://github.com/coreybutler/nvm-windows> (Windows)
