+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "7.0.109",
+    "version": "8.0.303",
     "rollForward": "latestMajor",
     "allowPrerelease": false
   }
