+++ nuget.config
@@ -1,14 +1,22 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
-  <packageSources>
-    <!--To inherit the global NuGet package sources remove the <clear/> line below -->
-    <clear />
-    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-  </packageSources>
-  <packageSourceMapping>
-    <!-- key value for <packageSource> should match key values from <packageSources> element -->
-    <packageSource key="nuget.org">
-      <package pattern="*" />
-    </packageSource>
-  </packageSourceMapping>
+	<packageSources>
+		<!--To inherit the global NuGet package sources remove the <clear/> line below -->
+		<clear />
+		<add key="LocalPackages" value="./nuget-local" />
+		<add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
+	</packageSources>
+	<packageSourceMapping>
+		<!-- key value for <packageSource> should match key values from <packageSources> element -->
+		<packageSource key="LocalPackages">
+			<!-- Uncomment to test local versions of the NuGet package -->
+			<!-- <package pattern="Kentico.Xperience.Contacts.Importer" /> -->
+
+			<!-- This entry doesn't match any packages but is required to satisfy the requirements of <packageSource> -->
+			<package pattern="LOCAL_PACKAGES_PLACEHOLDER" />
+		</packageSource>
+		<packageSource key="nuget.org">
+			<package pattern="*" />
+		</packageSource>
+	</packageSourceMapping>
 </configuration>
