+++ src/Kentico.Xperience.Contacts.Importer/Admin/Client/package.json
@@ -1,0 +1,41 @@
+{
+	"name": "kentico-xperience-integrations-contacts-import",
+	"version": "1.0.0",
+	"description": "Kentico Xperience Contacts Import project.",
+	"private": true,
+	"scripts": {
+		"start": "webpack serve --mode development",
+		"build": "webpack --mode=production",
+		"build:dev": "webpack --mode=development"
+	},
+	"engines": {
+		"node": ">=20.13 <=21"
+	},
+	"dependencies": {
+		"@kentico/xperience-admin-base": "29.2.2",
+		"@kentico/xperience-admin-components": "29.2.2",
+		"react": "18.3.1",
+		"react-dom": "18.3.1"
+	},
+	"devDependencies": {
+		"@babel/core": "7.24.9",
+		"@babel/plugin-transform-runtime": "7.24.7",
+		"@babel/preset-env": "7.24.8",
+		"@babel/preset-react": "7.24.7",
+		"@babel/preset-typescript": "7.24.7",
+		"@biomejs/biome": "^1.8.3",
+		"@kentico/xperience-webpack-config": "29.2.2",
+		"@types/react": "18.3.3",
+		"@types/react-dom": "18.3.0",
+		"babel-loader": "9.1.3",
+		"prettier": "3.3.3",
+		"webpack": "5.93.0",
+		"webpack-cli": "5.1.4",
+		"webpack-dev-server": "5.0.4"
+	},
+	"overrides": {
+		"react": "$react",
+		"react-dom": "$react"
+	},
+	"browserslist": ["> .2% and last 3 versions", "not op_mini all", "not IE 11"]
+}
