--- src/Kentico.Xperience.Contacts.Importer/Client/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "kentico-contacts-import-web-admin",
-  "version": "1.0.0",
-  "description": "Kentico Xperience Contacts Import project.",
-  "private": true,
-  "scripts": {
-    "start": "webpack serve --mode development",
-    "build": "webpack --mode=production",
-    "build:dev": "webpack --mode=development"
-  },
-  "engines": {
-    "node": ">=18.12.0 <=19"
-  },
-  "dependencies": {
-    "@kentico/xperience-admin-base": "26.3.1",
-    "@kentico/xperience-admin-components": "26.3.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
-  },
-  "devDependencies": {
-    "@babel/core": "7.22.9",
-    "@babel/plugin-transform-runtime": "7.22.9",
-    "@babel/preset-env": "7.22.9",
-    "@babel/preset-react": "7.22.5",
-    "@babel/preset-typescript": "7.22.5",
-    "@kentico/xperience-webpack-config": "26.3.1",
-    "@types/react": "18.2.18",
-    "@types/react-dom": "18.2.7",
-    "@typescript-eslint/eslint-plugin": "5.62.0",
-    "babel-loader": "9.1.3",
-    "eslint": "8.46.0",
-    "eslint-config-prettier": "8.9.0",
-    "eslint-config-standard-with-typescript": "37.0.0",
-    "eslint-plugin-import": "2.28.0",
-    "eslint-plugin-jsx-a11y": "6.7.1",
-    "eslint-plugin-n": "16.0.1",
-    "eslint-plugin-promise": "6.1.1",
-    "eslint-plugin-react": "7.33.1",
-    "eslint-plugin-react-hooks": "4.6.0",
-    "prettier": "3.0.0",
-    "webpack": "5.88.2",
-    "webpack-cli": "5.1.4",
-    "webpack-dev-server": "4.15.1"
-  },
-  "overrides": {
-    "react": "$react",
-    "react-dom": "$react"
-  },
-  "browserslist": [
-    "> .2% and last 3 versions",
-    "not op_mini all",
-    "not IE 11"
-  ]
-}
