+++ src/Kentico.Xperience.Contacts.Importer/ContactImportAdminModule.cs
@@ -3,7 +3,7 @@ using Kentico.Xperience.Contacts.Importer;
 
 [assembly: CMS.RegisterModule(typeof(ContactImportAdminModule))]
 [assembly: UICategory(
-   codeName: ContactImportAdminModule.CATEGORY,
+   codeName: "Kentico.Xperience.Contacts.Import.Web.Admin.Category",
    name: "Contacts Import",
    icon: Icons.PersonalisationVariants,
    order: 100)]
@@ -12,8 +12,6 @@ namespace Kentico.Xperience.Contacts.Importer;
 
 internal class ContactImportAdminModule : AdminModule
 {
-    public const string CATEGORY = "Kentico.Xperience.Contacts.Import.Web.Admin.Category";
-
     public ContactImportAdminModule()
         : base(nameof(ContactImportAdminModule))
     {
@@ -23,6 +21,6 @@ internal class ContactImportAdminModule : AdminModule
     {
         base.OnInit();
 
-        RegisterClientModule("kentico", "contacts-import-web-admin");
+        RegisterClientModule("kentico", "xperience-integrations-contacts-importer");
     }
 }
