+++ src/Kentico.Xperience.Contacts.Importer/Kentico.Xperience.Contacts.Importer.csproj
@@ -17,33 +17,37 @@
 		<LangVersion>latest</LangVersion>
 		<NoWarn>$(NoWarn);S1075</NoWarn>
 		<AdminOrgName>kentico</AdminOrgName>
+		<HotChocolateImplicitUsings>disable</HotChocolateImplicitUsings>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<AssemblyAttribute Include="CMS.AssemblyDiscoverableAttribute">
 		</AssemblyAttribute>
-		<AdminClientPath Include="Client\dist\**">
-			<ProjectName>contacts-import-web-admin</ProjectName>
+		<AdminClientPath Include="Admin\Client\dist\**">
+			<ProjectName>xperience-integrations-contacts-importer</ProjectName>
 		</AdminClientPath>
-		<Compile Remove="Client\node_modules\**" />
-		<EmbeddedResource Remove="Client\node_modules\**" />
-		<None Remove="Client\node_modules\**" />
+		<Compile Remove="Admin\Client\node_modules\**" />
+		<EmbeddedResource Remove="Admin\Client\node_modules\**" />
+		<None Remove="Admin\Client\node_modules\**" />
 	</ItemGroup>
 
 	<Target Name="NpmInstall" BeforeTargets="NpmBuildClient" Condition=" '$(Configuration)|$(Platform)' != 'Debug|AnyCPU' ">
-		<Exec Command="npm ci --no-audit --no-fund" WorkingDirectory="$(MSBuildProjectDirectory)\Client" />
+		<Exec Command="npm ci --no-audit --no-fund" WorkingDirectory="$(MSBuildProjectDirectory)\Admin\Client" />
 	</Target>
 	<Target Name="NpmBuildClient" BeforeTargets="BeforeBuild" Condition=" '$(Configuration)|$(Platform)' != 'Debug|AnyCPU' ">
-		<Exec Command="npm run build" WorkingDirectory="$(MSBuildProjectDirectory)\Client" />
+		<Exec Command="npm run build" WorkingDirectory="$(MSBuildProjectDirectory)\Admin\Client" />
+
+		<ItemGroup>
+			<AdminClientPath Include="Admin\Client\dist\**">
+				<ProjectName>xperience-integrations-contacts-importer</ProjectName>
+			</AdminClientPath>
+		</ItemGroup>
 	</Target>
 
+
 	<ItemGroup>
-		<PackageReference Include="CsvHelper" Version="30.0.1" />
-		<PackageReference Include="Kentico.Xperience.Admin" Version="26.3.2" />
-		<PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
+		<PackageReference Include="CsvHelper" />
+		<PackageReference Include="Kentico.Xperience.Admin" />
 	</ItemGroup>
 
 	<!--<ItemGroup>
