+++ test/Kentico.Xperience.Contacts.Importer.Generator/Kentico.Xperience.Contacts.Importer.Generator.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Bogus" Version="34.0.2" />
+    <PackageReference Include="Bogus" />
   </ItemGroup>
 
   <ItemGroup>
